🧹 refactor: cleaning up init event
- move Home methods into new module
- change some method names
- move Init event into new init method for Home
