+++ src/home/test_cases.rs
@@ -3,7 +3,7 @@ use crate::prelude::*;
 pub fn test_home() -> Home {
     let keymap = Config::new().unwrap().keymap.0;
 
-    let mut home = Home::new_builder();
+    let mut home = Home::builder();
 
     home.add_component(mock_items_table())
         .add_component(mock_receipts_table())
